# docs/specs/openapi.yaml
openapi: 3.1.0
info:
  title: Data Aggregator API
  version: 0.0.1
  description: Phase-one HTTP and SSE contract for Hybrid Chunk Search operations and Result Snapshots.
servers:
  - url: http://localhost:8080
security:
  - MockUserId: []
paths:
  /search-requests:
    post:
      summary: Create a Search Request
      description: Starts a Hybrid Chunk Search workflow and returns durable IDs plus the initial Operation state.
      operationId: createSearchRequest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchRequestCreateRequest"
            examples:
              hybridChunkSearch:
                value:
                  workflow: hybrid_chunk_search
                  keywords: digital wallet adoption
                  question: Which markets show the strongest payment growth?
                  retrieval_filters:
                    sector:
                      - financials
                    region:
                      - APAC
                      - Global
                  initial_sort:
                    field: relevance_score
                    direction: desc
      responses:
        "202":
          description: Search Request accepted and durable async work created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchRequestCreateResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"
  /operations/{operationId}:
    get:
      summary: Get an Operation
      operationId: getOperation
      parameters:
        - $ref: "#/components/parameters/OperationId"
      responses:
        "200":
          description: Current Operation state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operation"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
  /operations/{operationId}/cancel:
    post:
      summary: Cancel an Operation
      operationId: cancelOperation
      parameters:
        - $ref: "#/components/parameters/OperationId"
      responses:
        "200":
          description: Operation state after cancellation request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operation"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
  /events:
    get:
      summary: Subscribe to async Operation notifications
      operationId: subscribeEvents
      parameters:
        - name: scope_type
          in: query
          required: true
          schema:
            type: string
            enum:
              - result_snapshot
        - name: scope_id
          in: query
          required: true
          schema:
            type: string
          example: snap_123
      responses:
        "200":
          description: Server-sent events stream of lightweight Operation invalidation hints.
          content:
            text/event-stream:
              schema:
                type: string
              examples:
                progress:
                  value: |
                    event: async_run_progressed
                    data: {"operation_id":"op_123","type":"search_run","status":"running","scope_type":"result_snapshot","scope_id":"snap_123","warning_count":0}
        "400":
          description: Invalid scope parameters.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
            text/event-stream:
              schema:
                $ref: "#/components/schemas/EventStreamError"
        "404":
          description: The requested Result Snapshot event scope does not exist for the authenticated user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
            text/event-stream:
              schema:
                $ref: "#/components/schemas/EventStreamError"
        "500":
          description: Unexpected server failure.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
            text/event-stream:
              schema:
                $ref: "#/components/schemas/EventStreamError"
  /result-snapshots/{snapshotId}:
    get:
      summary: Get Result Snapshot metadata
      operationId: getResultSnapshot
      parameters:
        - $ref: "#/components/parameters/SnapshotId"
      responses:
        "200":
          description: Result Snapshot readiness metadata.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SnapshotMetadata"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
  /result-snapshots/{snapshotId}/activity:
    get:
      summary: Get Result Snapshot activity
      operationId: getResultSnapshotActivity
      parameters:
        - $ref: "#/components/parameters/SnapshotId"
      responses:
        "200":
          description: Operations associated with a Result Snapshot.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SnapshotActivity"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
  /result-snapshots/{snapshotId}/schema:
    get:
      summary: Get Result Snapshot schema
      operationId: getResultSnapshotSchema
      parameters:
        - $ref: "#/components/parameters/SnapshotId"
      responses:
        "200":
          description: Fields available for filtering, sorting, grouping, and aggregation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SnapshotSchema"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
  /result-snapshots/{snapshotId}/query:
    post:
      summary: Query a ready Result Snapshot
      operationId: queryResultSnapshot
      parameters:
        - $ref: "#/components/parameters/SnapshotId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SnapshotQueryRequest"
            examples:
              filteredGroupedQuery:
                value:
                  filters:
                    - field: sector
                      op: in
                      value:
                        - financials
                  sort:
                    - field: relevance_score
                      direction: desc
                      nulls: last
                  group_by:
                    - region
                  aggregations:
                    - name: result_count
                      op: count
                    - name: avg_relevance
                      op: avg
                      field: relevance_score
                  page:
                    limit: 10
                    offset: 0
      responses:
        "200":
          description: Query result rows, dynamic groups, and bounded page metadata.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SnapshotQueryResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Result Snapshot is not ready.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/ServerError"
  /internal/runtime:
    get:
      summary: Describe runtime mode
      operationId: describeRuntime
      security: []
      responses:
        "200":
          description: Internal runtime descriptor.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RuntimeDescriptor"
components:
  securitySchemes:
    MockUserId:
      type: apiKey
      in: header
      name: X-Mock-User-Id
      description: Phase-one mock user identity header.
  parameters:
    OperationId:
      name: operationId
      in: path
      required: true
      schema:
        type: string
      example: op_123
    SnapshotId:
      name: snapshotId
      in: path
      required: true
      schema:
        type: string
      example: snap_123
  responses:
    BadRequest:
      description: Invalid request body, query parameter, field, operator, sort, group, aggregation, or page bounds.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: The requested Operation or Result Snapshot does not exist for the authenticated user.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ServerError:
      description: Unexpected server failure.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    SearchRequestCreateRequest:
      type: object
      additionalProperties: false
      required:
        - workflow
        - keywords
        - question
        - retrieval_filters
        - initial_sort
      properties:
        workflow:
          type: string
          enum:
            - hybrid_chunk_search
        keywords:
          type: string
          minLength: 1
        question:
          type: string
          minLength: 1
        retrieval_filters:
          $ref: "#/components/schemas/SearchRequestRetrievalFilters"
        initial_sort:
          $ref: "#/components/schemas/SearchRequestInitialSort"
    SearchRequestRetrievalFilters:
      type: object
      additionalProperties: false
      description: At least one of sector, ticker, region, or published_at must be provided.
      properties:
        sector:
          type: array
          items:
            type: string
            minLength: 1
        ticker:
          type: array
          items:
            type: string
            minLength: 1
        region:
          type: array
          items:
            type: string
            minLength: 1
        published_at:
          $ref: "#/components/schemas/SearchRequestTimeRange"
    SearchRequestTimeRange:
      type: object
      additionalProperties: false
      description: At least one of from or to must be provided.
      properties:
        from:
          type: string
          format: date-time
        to:
          type: string
          format: date-time
    SearchRequestInitialSort:
      type: object
      additionalProperties: false
      required:
        - field
      properties:
        field:
          type: string
          enum:
            - relevance_score
            - published_at
            - lexical_rank
            - ticker
            - company_name
            - sector
            - region
        direction:
          type: string
          enum:
            - asc
            - desc
        nulls:
          type: string
          enum:
            - first
            - last
    SearchRequestCreateResponse:
      type: object
      additionalProperties: false
      required:
        - search_request_id
        - search_run_id
        - result_snapshot_id
        - operation
      properties:
        search_request_id:
          type: string
        search_run_id:
          type: string
        result_snapshot_id:
          type: string
        operation:
          $ref: "#/components/schemas/Operation"
    Operation:
      type: object
      additionalProperties: false
      required:
        - id
        - type
        - status
        - scope_type
        - scope_id
        - completed_steps
        - total_steps
        - warnings
        - created_at
        - updated_at
      properties:
        id:
          type: string
        type:
          type: string
          enum:
            - search_run
        status:
          $ref: "#/components/schemas/OperationStatus"
        scope_type:
          type: string
          enum:
            - result_snapshot
        scope_id:
          type: string
        current_step:
          type:
            - string
            - "null"
        completed_steps:
          type: integer
          minimum: 0
        total_steps:
          type: integer
          minimum: 0
        completed_units:
          type:
            - integer
            - "null"
          minimum: 0
        total_units:
          type:
            - integer
            - "null"
          minimum: 0
        warnings:
          type: array
          items:
            $ref: "#/components/schemas/OperationWarning"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    OperationStatus:
      type: string
      enum:
        - queued
        - running
        - waiting_retry
        - completed
        - completed_with_warnings
        - failed
        - cancelled
        - superseded
    OperationWarning:
      type: object
      additionalProperties: false
      properties:
        workflow_step_id:
          type:
            - string
            - "null"
        message:
          type:
            - string
            - "null"
    SnapshotMetadata:
      type: object
      additionalProperties: false
      required:
        - snapshot_id
        - search_run_id
        - status
        - created_at
      properties:
        snapshot_id:
          type: string
        search_run_id:
          type: string
        status:
          type: string
          enum:
            - pending
            - ready
            - cancelled
            - failed
        created_at:
          type: string
          format: date-time
        ready_at:
          type:
            - string
            - "null"
          format: date-time
    SnapshotActivity:
      type: object
      additionalProperties: false
      required:
        - scope_type
        - scope_id
        - operations
      properties:
        scope_type:
          type: string
          enum:
            - result_snapshot
        scope_id:
          type: string
        operations:
          type: array
          items:
            $ref: "#/components/schemas/Operation"
    SnapshotSchema:
      type: object
      additionalProperties: false
      required:
        - snapshot_id
        - fields
        - default_sort
      properties:
        snapshot_id:
          type: string
        fields:
          type: array
          items:
            $ref: "#/components/schemas/SnapshotSchemaField"
        default_sort:
          type: array
          items:
            $ref: "#/components/schemas/SnapshotSort"
    SnapshotSchemaField:
      type: object
      additionalProperties: false
      required:
        - name
        - type
        - filterable
        - sortable
        - groupable
        - aggregatable
        - nullable
      properties:
        name:
          type: string
        type:
          type: string
          enum:
            - string
            - enum
            - datetime
            - number
        filterable:
          type: boolean
        sortable:
          type: boolean
        groupable:
          type: boolean
        aggregatable:
          type: boolean
        nullable:
          type: boolean
    SnapshotQueryRequest:
      type: object
      additionalProperties: false
      properties:
        filters:
          type: array
          items:
            $ref: "#/components/schemas/SnapshotFilter"
        sort:
          type: array
          items:
            $ref: "#/components/schemas/SnapshotSort"
        group_by:
          type: array
          items:
            type: string
        aggregations:
          type: array
          items:
            $ref: "#/components/schemas/SnapshotAggregation"
        page:
          $ref: "#/components/schemas/SnapshotPageRequest"
    SnapshotFilter:
      type: object
      additionalProperties: false
      required:
        - field
        - op
        - value
      properties:
        field:
          type: string
        op:
          type: string
          enum:
            - in
            - eq
            - gt
            - gte
            - lt
            - lte
        value: {}
    SnapshotSort:
      type: object
      additionalProperties: false
      required:
        - field
      properties:
        field:
          type: string
        direction:
          type: string
          enum:
            - asc
            - desc
        nulls:
          type: string
          enum:
            - first
            - last
    SnapshotAggregation:
      type: object
      additionalProperties: false
      required:
        - name
        - op
      properties:
        name:
          type: string
          pattern: "^[a-z][a-z0-9_]*$"
        op:
          type: string
          enum:
            - count
            - avg
        field:
          type: string
    SnapshotPageRequest:
      type: object
      additionalProperties: false
      properties:
        limit:
          type: integer
          minimum: 1
          maximum: 500
          default: 50
        offset:
          type: integer
          minimum: 0
          default: 0
    SnapshotQueryResponse:
      type: object
      additionalProperties: false
      required:
        - snapshot_id
        - rows
        - groups
        - page
      properties:
        snapshot_id:
          type: string
        rows:
          type: array
          items:
            $ref: "#/components/schemas/SnapshotResultRow"
        groups:
          type: array
          items:
            type: object
            additionalProperties: true
            required:
              - key
            properties:
              key:
                type: object
                additionalProperties: true
          description: Dynamic group rows. Aggregation fields appear under the caller-provided aggregation names.
        page:
          $ref: "#/components/schemas/SnapshotPageResponse"
    SnapshotResultRow:
      type: object
      additionalProperties: false
      required:
        - chunk_id
        - parent_entity_id
        - parent_title
        - parent_type
        - source_name
        - ticker
        - company_name
        - sector
        - region
        - published_at
        - author
        - chunk_text
        - source_contributions
      properties:
        chunk_id:
          type: string
        parent_entity_id:
          type: string
        parent_title:
          type: string
        parent_type:
          type: string
        source_name:
          type: string
        ticker:
          type: string
        company_name:
          type: string
        sector:
          type: string
        region:
          type: string
        published_at:
          type: string
          format: date-time
        author:
          type: string
        chunk_text:
          type: string
        relevance_score:
          type:
            - number
            - "null"
        lexical_rank:
          type:
            - integer
            - "null"
        source_contributions:
          type: array
          items:
            type: string
    SnapshotPageResponse:
      type: object
      additionalProperties: false
      required:
        - limit
        - offset
        - total
      properties:
        limit:
          type: integer
          minimum: 1
          maximum: 500
        offset:
          type: integer
          minimum: 0
        total:
          type: integer
          minimum: 0
    RuntimeDescriptor:
      type: object
      additionalProperties: false
      required:
        - application
        - runtimeMode
        - mockUserId
        - workflowConfigVersion
      properties:
        application:
          type: string
        runtimeMode:
          type: string
        mockUserId:
          type: string
        workflowConfigVersion:
          type: integer
    ErrorResponse:
      type: object
      additionalProperties: false
      description: Stable error response for validation, not-found, conflict, and server failures.
      required:
        - code
        - message
        - status
        - path
        - details
      properties:
        code:
          type: string
          examples:
            - VALIDATION_ERROR
            - BAD_REQUEST
            - NOT_FOUND
            - CONFLICT
        message:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        path:
          type: string
        details:
          type: object
          additionalProperties: true
    EventStreamError:
      type: string
      description: SSE error event whose data payload is an ErrorResponse JSON object.
      examples:
        - |
          event: error
          data: {"code":"NOT_FOUND","message":"Unknown result snapshot: snap_123","status":404,"path":"/events","details":{}}